requirements.txt
Django
django-crispy-forms
djangorestframework
djoser